requirements.txt
amqp==5.0.1
appnope==0.1.2
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
billiard==3.6.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-author==1.0.2
django-ckeditor==6.0.0
django-cleanup==5.0.0
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-import-export==2.3.0
django-js-asset==1.2.2
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.1
jdcal==1.4.1
jedi==0.17.2
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.24
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2-binary==2.9.1
ptyprocess==0.6.0
Pygments==2.7.1
pylint==2.6.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
setuptools-git==1.2
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
traitlets==5.0.4
urllib3==1.26.3
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
